requirements.txt
streamlit==1.28.0
pymongo==4.6.0
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0
